// src/modules/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  // Auth & Account
  ACCOUNT_VERIFIED = 'ACCOUNT_VERIFIED',
  ACCOUNT_SUSPENDED = 'ACCOUNT_SUSPENDED',

  // Bids & Proposals
  NEW_BID_RECEIVED = 'NEW_BID_RECEIVED',
  BID_ACCEPTED = 'BID_ACCEPTED',
  BID_REJECTED = 'BID_REJECTED',
  BID_WITHDRAWN = 'BID_WITHDRAWN',

  // Projects
  PROJECT_CREATED = 'PROJECT_CREATED',
  PROJECT_ASSIGNED = 'PROJECT_ASSIGNED',
  PROJECT_UPDATED = 'PROJECT_UPDATED',
  PROJECT_COMPLETED = 'PROJECT_COMPLETED',

  // Milestones
  MILESTONE_CREATED = 'MILESTONE_CREATED',
  MILESTONE_STARTED = 'MILESTONE_STARTED',
  MILESTONE_IN_PROGRESS = 'MILESTONE_IN_PROGRESS',
  MILESTONE_SUBMITTED = 'MILESTONE_SUBMITTED',
  MILESTONE_APPROVED = 'MILESTONE_APPROVED',
  MILESTONE_REJECTED = 'MILESTONE_REJECTED',

  // Payments
  PAYMENT_INITIATED = 'PAYMENT_INITIATED',
  PAYMENT_COMPLETED = 'PAYMENT_COMPLETED',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  PAYOUT_SENT = 'PAYOUT_SENT',
  PAYOUT_RECEIVED = 'PAYOUT_RECEIVED',

  // Messages
  NEW_MESSAGE = 'NEW_MESSAGE',

  // System
  SYSTEM_ALERT = 'SYSTEM_ALERT',
}